Split tool config into local and remote
