+++ notifier/__main__.py
@@ -4,7 +4,7 @@ import keyring
 from apscheduler.schedulers.blocking import BlockingScheduler
 from apscheduler.triggers.cron import CronTrigger
 
-from notifier.config.tool import read_local_config
+from notifier.config.local import read_local_config
 from notifier.database.drivers import DatabaseDriver
 from notifier.tasks import notification_channels, notify_active_channels
 from notifier.types import LocalConfig
